Deleted the uncessaryfiles and created NavigationLayout
