+++ src/routes/Account.tsx
@@ -1,10 +1,8 @@
-import { NavigationBar } from "../components/NavigationBar";
 export function Account() {
   return (
     <div>
       <h1>Login</h1>
       <h1>Account</h1>
-      <NavigationBar />
     </div>
   );
 }
