+++ src/routes/Bookmark.tsx
@@ -1,10 +1,7 @@
-import { NavigationBar } from "../components/NavigationBar";
-
 export function Bookmark() {
   return (
     <div>
       <h1>Bookmark</h1>
-      <NavigationBar />
     </div>
   );
 }
