+++ src/routes/Home.tsx
@@ -1,9 +1,7 @@
-import { NavigationBar } from "../components/NavigationBar";
 export function Home() {
   return (
     <div>
       <h1>Home</h1>
-      <NavigationBar />
     </div>
   );
 }
