+++ src/routes/MoviesList.tsx
@@ -1,10 +1,7 @@
-import { NavigationBar } from "../components/NavigationBar";
-
 export function MoviesList() {
   return (
     <div>
       <h1>Movies</h1>
-      <NavigationBar />
     </div>
   );
 }
